# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Consulta os dados do usuário autenticado
      description: Retorna perfil, empresa, permissões e status da conta vinculada ao Bearer Token JWT.
      operationId: getCurrentUser
      responses:
        '200':
          description: Usuário autenticado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: '00000000-0000-0000-0000-000000000001'
                username: admin@autocarehub.com
                role: ADMIN
                profileType: MASTER_ADMIN
                fullName: Marina AutoCare Hub
                permissions:
                  - VIEW_STATS
                active: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Users
      summary: Atualiza os dados básicos do usuário autenticado
      description: Permite ao usuário logado alterar o próprio nome exibido no sistema.
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserRequest'
            example:
              fullName: Maria AutoCare Hub
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/password:
    patch:
      tags:
        - Users
      summary: Altera a senha do usuário autenticado
      description: Valida a senha atual e define uma nova senha para a conta logada.
      operationId: changeCurrentPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: 'autocare123'
              newPassword: novaSenha123
      responses:
        '204':
          description: Senha alterada
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/preferences/home:
    get:
      tags:
        - Users
      summary: Consulta preferências da home do usuário
      description: Retorna os widgets visíveis e a preferência de exibir alertas na home.
      operationId: getHomePreference
      responses:
        '200':
          description: Preferência encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomePreferenceResponse'
              example:
                widgets:
                  - orders-progress
                  - services-catalog
                showAlertsOnHome: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Users
      summary: Salva as preferências da home do usuário
      description: Persiste no backend os widgets escolhidos para a home do usuário autenticado.
      operationId: saveHomePreference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomePreferenceRequest'
            example:
              widgets:
                - orders-progress
                - active-customers
                - vehicles-in-service
              showAlertsOnHome: true
      responses:
        '200':
          description: Preferência salva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomePreferenceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Lista contas de usuário
      description: >
        Endpoint administrativo para listar usuários por status, role, perfil ou termo de busca.
        Requer Bearer Token JWT de administrador.
      operationId: listUsers
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: boolean
        - name: role
          in: query
          required: false
          schema:
            type: string
            example: ADMIN
        - name: profileType
          in: query
          required: false
          schema:
            type: string
            example: WORKSHOP_ADMIN
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
      responses:
        '200':
          description: Usuários encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Users
      summary: Cria uma conta de usuário
      description: >
        Endpoint administrativo para criar administradores, funcionários ou clientes vinculados ao sistema.
        Requer Bearer Token JWT de administrador. IDs de usuário e de empresa nova são gerados pelo backend.
        Para vincular uma conta a empresa existente, consulte antes `GET /api/v1/users/companies` e use a empresa selecionada.
        Para criar uma empresa nova, envie `createCompany: true`, `companyName` e `companyType`, sem informar `companyId`.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              funcionarioEmpresaExistente:
                summary: Funcionário em empresa existente selecionada
                value:
                  username: mecanico@example.com
                  password: 'autocare123'
                  role: EMPLOYEE
                  fullName: Carlos Oficina
                  profileType: WORKSHOP_EMPLOYEE
                  companyId: 90000000-0000-0000-0000-000000000011
                  companyName: Oficina Central AutoCare
                  companyType: WORKSHOP
                  createCompany: false
                  employeeSubRole: MECHANIC
                  permissions:
                    - CREATE_ORDER
                    - EDIT_ORDER
                  active: true
              adminNovaEmpresa:
                summary: Admin Master criando nova oficina sem informar id
                value:
                  username: nova.oficina@example.com
                  password: 'autocare123'
                  role: ADMIN
                  fullName: Ana Oficina
                  profileType: WORKSHOP_ADMIN
                  companyName: Oficina Nova AutoCare
                  companyType: WORKSHOP
                  createCompany: true
                  employeeSubRole: ''
                  permissions:
                    - VIEW_BILLING
                    - MANAGE_STOCK
                    - CREATE_BUDGET
                    - EDIT_EMPLOYEES
                    - VIEW_STATS
                  active: true
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/partners:
    get:
      tags:
        - Users
      summary: Lista parceiros ativos
      description: Lista oficinas e lojas de peças cadastradas como parceiros ativos para uso na area do cliente final.
      operationId: listPartners
      responses:
        '200':
          description: Parceiros encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/companies:
    get:
      tags:
        - Users
      summary: Lista empresas disponíveis para vínculo de usuários
      description: >
        Endpoint administrativo usado no cadastro de contas. Admin Master visualiza todas as empresas.
        Admin de oficina ou loja visualiza somente a própria empresa.
      operationId: listCompaniesForUsers
      responses:
        '200':
          description: Empresas encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListResponse'
              example:
                items:
                  - id: 90000000-0000-0000-0000-000000000001
                    name: AutoCare Hub
                    type: PLATFORM
                    active: true
                  - id: 90000000-0000-0000-0000-000000000011
                    name: Oficina Central AutoCare
                    type: WORKSHOP
                    active: true
                  - id: 90000000-0000-0000-0000-000000000012
                    name: Loja peças Prime
                    type: PARTS_STORE
                    active: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/{userId}:
    put:
      tags:
        - Users
      summary: Atualiza uma conta de usuário
      description: Endpoint administrativo para editar dados, perfil, subrole, permissões e status ativo de uma conta.
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/{userId}/password:
    patch:
      tags:
        - Users
      summary: Redefine a senha de uma conta
      description: >
        Endpoint administrativo para definir nova senha inicial ou resetar senha de funcionário, cliente ou parceiro.
      operationId: resetUserPassword
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Senha redefinida
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
    Forbidden:
      description: Usuário autenticado sem permissão para acessar o recurso, ou token externo ausente/invalido nos webhooks simulados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 403
            error: Forbidden
            message: Access denied
            path: /api/v1/customers
            details: [ ]
...
    Conflict:
      description: Conflito com estado atual do recurso, como duplicidade de documento, placa, usuário ou SKU
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 409
            error: Conflict
            message: Resource already exists
            path: /api/v1/customers
            details: [ ]
    InternalServerError:
      description: Erro interno inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 500
            error: Internal Server Error
            message: Unexpected error
            path: /api/v1/service-orders
            details: [ ]
  schemas:
...
    UserResponse:
      type: object
      required:
        - id
        - username
        - role
        - fullName
        - profileType
        - permissions
        - active
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          format: email
          maxLength: 120
        role:
          type: string
          example: ADMIN
        customerId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
          description: Use somente para vincular a uma empresa existente retornada por GET /api/v1/users/companies. Não informe ao criar uma nova empresa.
        fullName:
          type: string
          maxLength: 160
        profileType:
          type: string
          example: WORKSHOP_ADMIN
        companyName:
          type: string
          maxLength: 160
        companyType:
          type: string
          example: workshop
        employeeSubRole:
          type: string
          example: MECHANIC
        permissions:
          type: array
          items:
            type: string
          example:
            - CREATE_ORDER
            - EDIT_ORDER
            - MANAGE_STOCK
        active:
          type: boolean
...
    CompanyListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompanyResponse'
    UserListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - role
        - fullName
        - profileType
        - active
      properties:
        username:
          type: string
          format: email
          maxLength: 120
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
        role:
          type: string
          enum:
            - ADMIN
            - EMPLOYEE
            - CUSTOMER
        customerId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        fullName:
          type: string
          minLength: 1
          maxLength: 160
        profileType:
          type: string
          example: WORKSHOP_EMPLOYEE
        companyName:
          type: string
          maxLength: 160
        companyType:
          type: string
          enum:
            - PLATFORM
            - WORKSHOP
            - PARTS_STORE
          example: WORKSHOP
        createCompany:
          type: boolean
          default: false
          description: Quando true, o backend cria a empresa e gera o companyId automaticamente.
        employeeSubRole:
          type: string
          example: MECHANIC
        permissions:
          type: array
          items:
            type: string
        active:
          type: boolean
...
    UpdateCurrentUserRequest:
      type: object
      required:
        - fullName
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 160
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 1
          maxLength: 72
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 72
    ResetPasswordRequest:
      type: object
      required:
        - newPassword
      properties:
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 72
    HomePreferenceRequest:
      type: object
      required:
        - widgets
        - showAlertsOnHome
      properties:
        widgets:
          type: array
          minItems: 1
          items:
            type: string
        showAlertsOnHome:
          type: boolean
    HomePreferenceResponse:
      allOf:
        - $ref: '#/components/schemas/HomePreferenceRequest'
...